Introduce ledger storage adapter seam
