+++ packages/ledger/src/archive.ts
@@ -37,6 +37,7 @@ installSqliteWarningFilter();
 
 import type { TurnRecord } from '@relayburn/reader';
 
+import { getStorageAdapterKind } from './adapters/factory.js';
 import { withLock } from './lock.js';
 import { archivePath, ledgerPath } from './paths.js';
 import {
@@ -311,6 +312,18 @@ export interface BuildResult {
  * we delete and recreate, since the archive is by design rebuildable.
  */
 export async function openArchive(): Promise<DatabaseSync> {
+  if (!archiveEnabled()) {
+    const db = new DatabaseSync(':memory:');
+    db.exec('PRAGMA foreign_keys = ON');
+    db.exec(SCHEMA_SQL);
+    applyAdditiveMigrations(db);
+    db.prepare(
+      `INSERT INTO archive_state (id, ledger_offset_bytes, ledger_mtime_ms, archive_version)
+       VALUES (1, 0, 0, ?)`,
+    ).run(ARCHIVE_VERSION);
+    return db;
+  }
+
   const dbPath = archivePath();
   await mkdir(path.dirname(dbPath), { recursive: true });
 
@@ -400,6 +413,49 @@ async function fileExists(p: string): Promise<boolean> {
   }
 }
 
+function archiveEnabled(): boolean {
+  return getStorageAdapterKind() === 'file';
+}
+
+function emptyBuildResult(): BuildResult {
+  return {
+    scannedBytes: 0,
+    turnsApplied: 0,
+    sessionsTouched: 0,
+    stampsApplied: 0,
+    compactionsApplied: 0,
+    toolResultEventsApplied: 0,
+    rebuiltFromZero: false,
+  };
+}
+
+function emptyArchiveStatus(
+  archiveFilePath: string,
+  ledgerSizeBytes: number,
+  ledgerMtimeMsCurrent: number,
+): ArchiveStatus {
+  return {
+    archivePath: archiveFilePath,
+    exists: false,
+    archiveVersion: ARCHIVE_VERSION,
+    ledgerOffsetBytes: 0,
+    ledgerMtimeMs: 0,
+    ledgerSizeBytes,
+    ledgerMtimeMsCurrent,
+    upToDate: false,
+    lastBuiltAt: null,
+    lastRebuildAt: null,
+    rowCounts: {
+      sessions: 0,
+      turns: 0,
+      toolCalls: 0,
+      toolResultEvents: 0,
+      compactions: 0,
+    },
+    fidelityHistogram: {},
+  };
+}
+
 /**
  * Drop the archive entirely and rebuild from the ledger.
  *
@@ -409,6 +465,7 @@ async function fileExists(p: string): Promise<boolean> {
  * threw).
  */
 export async function rebuildArchive(): Promise<BuildResult> {
+  if (!archiveEnabled()) return emptyBuildResult();
   return withLock('archive', async () => {
     const dbPath = archivePath();
     await unlink(dbPath).catch(() => undefined);
@@ -431,6 +488,7 @@ export async function rebuildArchive(): Promise<BuildResult> {
  * path.
  */
 export async function buildArchive(): Promise<BuildResult> {
+  if (!archiveEnabled()) return emptyBuildResult();
   return withLock('archive', () => buildArchiveLocked());
 }
 
@@ -454,6 +512,15 @@ export async function buildArchive(): Promise<BuildResult> {
  * against an artificially small pre-vacuum number.
  */
 export async function vacuumArchive(): Promise<VacuumResult> {
+  if (!archiveEnabled()) {
+    return {
+      archivePath: archivePath(),
+      existed: false,
+      beforeBytes: 0,
+      afterBytes: 0,
+      reclaimedBytes: 0,
+    };
+  }
   return withLock('archive', async () => {
     const dbPath = archivePath();
     if (!(await fileExists(dbPath))) {
@@ -1244,6 +1311,9 @@ async function* readJsonlFrom(filePath: string, startOffset: number): AsyncItera
  * consistent read view.
  */
 export async function getArchiveStatus(): Promise<ArchiveStatus> {
+  if (!archiveEnabled()) {
+    return emptyArchiveStatus(archivePath(), 0, 0);
+  }
   const dbPath = archivePath();
   const exists = await fileExists(dbPath);
   const ledger = ledgerPath();
